Add acceptance spec - asking users to sign in
This commit also renames the other acceptance spec. Notably, we're
putting these under 'feature' specs in Rails terminology, but we're
using them like ATDD acceptance specs.
