+++ app/controllers/monitoring_controller.rb
@@ -1,4 +1,6 @@
 class MonitoringController < ApplicationController
+  skip_before_action :authenticate_user!
+
   def healthcheck
     render body: "Healthy"
   end
